Added uint32_t ssaID to DXIL::GlobalVar
Changed GlobalVar::align from uint64_t -> uint32_t to make space for the new member

+++ renderdoc/driver/shaders/dxil/dxil_bytecode_editor.cpp
@@ -598,7 +598,7 @@ bytebuf ProgramEditor::EncodeProgram()
 
   for(size_t i = 0; i < m_GlobalVars.size(); i++)
   {
-    cfg.maxAlign = RDCMAX(m_GlobalVars[i]->align, cfg.maxAlign);
+    cfg.maxAlign = RDCMAX((uint64_t)m_GlobalVars[i]->align, cfg.maxAlign);
     RDCASSERT(m_GlobalVars[i]->type->type == Type::Pointer);
     uint32_t typeIndex = uint32_t(getTypeID(m_GlobalVars[i]->type->inner));
     cfg.maxGlobalType = RDCMAX(typeIndex, cfg.maxGlobalType);
